Makefile: Adjust for building out-of-tree
In order to call the Makefile from another directory using the `make -C`
flag minor adjustments are necessary.

Especially we cannot rely on the `PWD` variable anymore and should
define the drivers root directory by the current Makefile location.
A common way is leveraging the MAKEFILE_LIST and figure out the dirname.

On the way make it possible to set the KERNELDIR variable from outside
the Makefile by using the `?=` assignment instead of the `=`.

Also, remove the unused ROOT_DIR from the app build command.